ok somewhat faster... less than 4 minutes

+++ util/src/main/java/org/fde/util/primes/sieve/store/CalculatePrime.java
@@ -2,16 +2,16 @@ package org.fde.util.primes.sieve.store;
 
 public class CalculatePrime {
     private final Store store;
-    private final int maximum;
+    private final long maximum;
 
     public CalculatePrime(Store store) {
         this.store = store;
         int maxPrime = new MaxPrime(store).getMaxPrime();
-        this.maximum = maxPrime * maxPrime;
+        this.maximum = ((long)maxPrime) * maxPrime;
     }
 
     public boolean isCalculatedPrime(long suspect) {
-        if(suspect > this.maximum) {
+        if (suspect > this.maximum) {
             String msg = String.format("suspect (%s) > maximum (%s)",
                     suspect, this.maximum);
 
@@ -20,11 +20,10 @@ public class CalculatePrime {
 
         int length = store.getLength();
 
-        if(suspect < length) {
+        if (suspect < length) {
             boolean isPrime = store.isPrime((int) suspect);
             return isPrime;
-        }
-        else {
+        } else {
             long limit = Math.round(Math.sqrt(suspect) + 1);
 
             for (int index = 2; index < limit; ++index) {
@@ -32,7 +31,7 @@ public class CalculatePrime {
                     ++index;
                 }
 
-                if(index >= length) {
+                if (index >= length) {
                     String msg = String.format("index (%s) > store length (%s)",
                             index, length);
